+++ posthog/warehouse/models/snapshot_job.py
@@ -14,7 +14,7 @@ class DataWarehouseSnapshotJob(CreatedMetaFields, UpdatedMetaFields, UUIDTModel,
     config = models.ForeignKey("posthog.DataWarehouseSnapshotConfig", on_delete=models.CASCADE)
 
     # Each snapshot job will create a saved query that is filtered to the state of the table at run time
-    table = models.ForeignKey("posthog.DataWarehouseSavedQuery", on_delete=models.CASCADE, null=True, blank=True)
+    saved_query = models.ForeignKey("posthog.DataWarehouseSavedQuery", on_delete=models.CASCADE, null=True, blank=True)
     status = models.CharField(max_length=400, choices=Status.choices, default=Status.RUNNING)
     error = models.TextField(null=True, blank=True)
     workflow_id = models.CharField(max_length=400, null=True, blank=True)
